+++ src/hexconvert.h
@@ -1,0 +1,9 @@
+/*Copyright (c) Tristan Wellman*/
+#ifndef HEXCONVERT_H
+#define HEXCONVERT_H
+#include <stdint.h>
+
+char *uint64ToHex(char *value);
+char *float64ToHex(char *value); 
+
+#endif